pyproject.toml
[project]
name = "abm-agricultural-extension"
version = "0.1.0"
description = "An agent-based model of agricultural extension"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "colorcet>=3.1.0",
    "geopandas>=1.0.1",
    "hydra-core>=1.3.2",
    "hydra-joblib-launcher>=1.2.0",
    "ipykernel>=6.29.5",
    "matplotlib>=3.10.0",
    "networkx>=3.4.2",
    "omegaconf>=2.3.0",
    "pyarrow>=19.0.1",
    "scikit-learn>=1.6.1",
    "scipy>=1.15.2",
    "shortuuid>=1.0.13",
    "sortedcontainers>=2.4.0",
]
